Moved painter.js to src/utils
